Added iodine server + minor updates
I’ve added the iodine server for Ruby MRI.

I’ve also removed the reference to Sinatra beta, since version 2 was
released.

I updated `run.sh` to remove the benchmark result folders when starting
a new benchmark.

This will both prevent contamination (just in case) and automate the
folder tree to be match any updates in the server list.

In `setup.rb` I’ve moved things around just to make sure that an
exception isn’t thrown on `drop_table`. The script seemed to assume a
prior database existed and this change simply protects against cases in
which this assumption might be incorrect.

It was a joy to run the benchmarks on my machine.

At the moment I fear the use of the server’s default concurrency model
might give some servers (i.g,`iodine` and `puma`) an unfair advantage
related to core-matching defaults… but I couldn’t find anything to do
about this.